+++ packages/typespec-python/test/unbranded/mock_api_tests/test_unbranded.py
@@ -54,4 +54,9 @@ def test_sensitive_word():
     check_folder = (Path(os.path.dirname(__file__)) / "../generated").resolve()
     assert [] == check_sensitive_word(check_folder, "azure")
     # after update cadl-ranch, it shall also equal to []
-    assert ["authentication-oauth2", "authentication-union"] == check_sensitive_word(check_folder, "microsoft")
+    assert [
+        "authentication-oauth2",
+        "authentication-union",
+        "generation-subdir",
+        "setuppy-authentication-union",
+    ] == check_sensitive_word(check_folder, "microsoft")